feat: extract instruction data prep and rendering
